# deliverables_sprint1_2/openapi.yaml
openapi: 3.0.0
info:
  title: Student Club Management API
  description: >
    API documentation for the Student Club Management Application. 
    
    ### APIs Integrated:
    - Google Gemini AI API (@google/genai SDK): Used for `/api/gemini/summarize-meeting`, `/api/gemini/task-breakdown`, and `/api/gemini/event-planner`.
    
    ### APIs Created:
    - Auth APIs (`/api/auth/register`, `/api/auth/login`, `/api/auth/me`)
    - User Management APIs (`/api/users`, `/api/users/{id}/role`)
    - Admin Stats API (`/api/admin/stats`)
    - Announcements APIs (`/api/announcements` GET/POST/PUT/DELETE)
    - Tasks APIs (`/api/tasks` GET/POST)
    - Health API (`/api/health`)

    ### User Stories Mapping:
    - **US-1**: "As a new user, I want to register and login so that I can access the system securely." (Maps to Auth APIs, handles duplicate email errors (400) and invalid credentials (401))
    - **US-2**: "As an admin (Club Head), I want to see statistics of the platform and manage member roles." (Maps to Admin Stats API, User listing, and Role update API. Handles permission errors (403))
    - **US-3**: "As a core member, I want to manage tasks and announce events." (Maps to Tasks and Announcements APIs)
    - **US-4**: "As a user, I want to use AI to generate event plans, summarize meetings, and breakdown tasks." (Maps to Gemini AI endpoints)
    
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Dev Server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      
security:
  - bearerAuth: []

paths:
  /api/auth/register:
    post:
      security: []
      summary: Register User
      description: Register a new user and return a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, email, password]
              properties:
                fullName: { type: string }
                email: { type: string }
                password: { type: string }
                department: { type: string }
                academicYear: { type: integer }
      responses:
        '201':
          description: User created successfully
        '400':
          description: Missing fields or Email already exists
        '500':
          description: Server Error

  /api/auth/login:
    post:
      security: []
      summary: Login User
      description: Authenticate user and return a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Successfully authenticated
        '400':
          description: Missing email or password
        '401':
          description: Invalid credentials
        '500':
          description: Server Error

  /api/auth/me:
    get:
      summary: Get Current User
      description: Fetch the current user details using JWT token.
      responses:
        '200':
          description: Current user details
        '401':
          description: No token provided or Invalid token
        '404':
          description: User not found

  /api/users:
    get:
      summary: Get All Users
      description: Fetch all users. Requires authentication.
      responses:
        '200':
          description: List of users
        '401':
          description: Unauthorized
        '500':
          description: Server Error

  /api/users/{id}/role:
    put:
      summary: Update User Role
      description: Update a user's role. Only accessible to Admin or Club Head.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string }
      responses:
        '200':
          description: Role updated successfully
        '400':
          description: Role is required
        '401':
          description: Unauthorized
        '403':
          description: Only Club Head or Admin can change roles
        '500':
          description: Server Error

  /api/admin/stats:
    get:
      summary: Get Admin Dashboard Stats
      description: Fetch aggregate statistics for the admin dashboard.
      responses:
        '200':
          description: Statistics object
        '401':
          description: Unauthorized
        '403':
          description: Access denied (Not Admin)
        '500':
          description: Server Error

  /api/health:
    get:
      security: []
      summary: Health Check
      description: Basic server health check.
      responses:
        '200':
          description: OK